requirements.txt
Flask==2.3.2
PyPDF2==3.1.1
python-docx==0.8.11
spacy==3.7.0